requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
distro==1.9.0
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.0
griffe==1.6.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.21
langchain-core==0.3.49
langchain-text-splitters==0.3.7
langsmith==0.3.19
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.5.0
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
openai==1.68.2
openai-agents==0.0.7
orjson==3.10.16
packaging==24.2
pillow==11.1.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
safetensors==0.5.3
setuptools==78.1.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.39
sse-starlette==2.2.1
starlette==0.46.1
sympy==1.13.1
tenacity==9.0.0
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.50.2
typer==0.15.2
types-requests==2.32.0.20250306
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
zstandard==0.23.0
